+++ taskgraph/__init__.py
@@ -1,9 +1,8 @@
 """TaskGraph init module."""
-from pkg_resources import get_distribution
 
 from .Task import TaskGraph
 from .Task import Task
 from .Task import _TASKGRAPH_DATABASE_FILENAME
+from .Task import __version__
 
-__all__ = ['TaskGraph', 'Task', '_TASKGRAPH_DATABASE_FILENAME']
-__version__ = get_distribution(__name__).version
+__all__ = ['__version__', 'TaskGraph', 'Task', '_TASKGRAPH_DATABASE_FILENAME']